temporarily add pycomlink as submodule
I do this so that I can run the copy-pasted code with the OpenMRG example data here, visually check the results and then adjust the test that rely on DWD data to use the OpenMRG data. Only then I will change the code to use polirgain instaead of wradlib and pycomlink code.