remove drop_vars('time') which I added to fix some problem in one of last PRs
But I do not recall why exactly. It causese issue when passing data that has no time dimension and also no variable or coord named `time`, which happens with my synthetic example data. Maybe I found the problem when working in GridAtPoints and GridAtLines in poligrain and then just added the same code here without needing it. That means, thouhg, that the same problem might appear in GridAtLines and GridAtPoints. But let's wait and see if this really pops up as problem in poligrian...